Implement basic routing with react-router-dom
